--- a/src/Makefile.am
+++ b/src/Makefile.am
@@ -19,6 +19,9 @@ sources.cpp step.cpp step_db.cpp stress.cpp structure.cpp structure_dump.cpp		\
 susceptibility.cpp time.cpp update_eh.cpp mpb.cpp update_pols.cpp 	\
 vec.cpp step_generic.cpp meepgeom.cpp GDSIIgeom.cpp $(HDRS) $(BUILT_SOURCES)
 
+SUBDIRS = support
+libmeep_la_LIBADD = support/libsupport.la
+
 libmeep_la_LDFLAGS = -version-info @SHARED_VERSION_INFO@
 
 PRELUDE = "/* This file was automatically generated --- DO NOT EDIT */"