--- a/src/support/Makefile.am
+++ b/src/support/Makefile.am
@@ -1,0 +1,3 @@
+noinst_LTLIBRARIES = libsupport.la
+
+libsupport_la_SOURCES = mt19937ar.c meep_mt.h