Now defaults to Mersenne Twister random-number generation (included under 3-clause BSD license) and defaulting to POSIX Bessel functions rather than the much-slower GSL ones for near2far calculation (for #861).