@stevengj 2019-05-09
@oskooi, try the near2far calculation with this PR — it should be much faster (in 2d) with the new Bessel functions.