@oskooi 2019-05-10
The following are the wall-clock `Field time usage` statistics for `examples/binary_grating_n2f.py` using `resolution=25`, `nfreq=50`, and `nperiods=40` with MPI (`np=2`) comparing master and this PR: 

**master**

*unit cell with periodic boundaries*
```
Field time usage:
        connecting chunks: 0.00340685 s
            time stepping: 1.70402 s
            communicating: 0.368193 s
     Fourier transforming: 0.0024059 s
        getting farfields: 1525.26 s
          everything else: 2.16774 s
```
*finite periodic grating with flat surface termination extending into PML*
```
Field time usage:
	connecting chunks: 0.179651 s
            time stepping: 270.982 s
            communicating: 9.90172 s
     Fourier transforming: 125.645 s
	getting farfields: 763.141 s
          everything else: 2.1376 s
```

**this PR**

*unit cell with periodic boundaries*
```
Field time usage:
	connecting chunks: 0.00318903 s
            time stepping: 1.71046 s
            communicating: 0.369378 s
     Fourier transforming: 0.00197067 s
        getting farfields: 843.431 s
          everything else: 2.1393 s
```
*finite periodic grating with flat surface termination extending into PML*
```
Field time usage:
	connecting chunks: 0.187708 s
            time stepping: 272.32 s
            communicating: 9.89654 s
     Fourier transforming: 126.648 s
	getting farfields: 420.911 s
          everything else: 2.61026 s
```

The speedup in `get_farfields` is significant as the runtime is nearly *halved* for both the unit- and super-cell calculations: `1525.26 s` vs. `843.431 s` and `763.141 s` vs. `420.911 s`.